Bump MSRV to 1.42
This means that we will support approximately 2 years of old rust
versions, which I think is a fairly reasonable MSRV policy.
