+++ Cargo.toml
@@ -26,6 +26,7 @@ backtrace = { version = "0.3", optional = true }
 
 [dev-dependencies]
 procinfo = "0.4.2"
+failure = "0.1"
 
 [package.metadata.docs.rs]
 all-features = true
