--- procfs-tests/Cargo.toml
@@ -1,13 +1,0 @@
-[package]
-name = "procfs-tests"
-version = "0.1.0"
-authors = ["Andrew Chin <achin@eminence32.net>"]
-edition = "2018"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-procfs = {path = ".."}
-
-[dev-dependencies]
-failure = "0.1"
